requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
httpx>=0.25.0
pandas>=2.0.0
openai>=1.3.0
